refactor cmd package
