+++ cmd/stop.go
@@ -1,0 +1,23 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/yohamta/dagu/internal/controller"
+	"log"
+)
+
+func stopCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "stop <DAG file>",
+		Short: "Stop the running DAG",
+		Long:  `dagu stop <DAG file>`,
+		Args:  cobra.ExactArgs(1),
+		Run: func(cmd *cobra.Command, args []string) {
+			loadedDAG, err := loadDAG(args[0], "")
+			checkError(err)
+
+			log.Printf("Stopping...")
+			checkError(controller.NewDAGController(loadedDAG).Stop())
+		},
+	}
+}
